+++ NEWS.md
@@ -2,6 +2,7 @@
 * fixed plot 2020 03 18 intensive veneto
 * 20200325 modul veneto hosp
 * fix colour focus_picco (#24)
+* translated 20200323 picco (#23)
 
 # covid19ita 0.19.0